Create CSIT102_PYJACK
README
Write a Python program to simulate a single round of a fictional game Pyjack (a variation of Blackjack). You can find the rules of Blackjack explained here. For Pyjack, the rules are as follows:
The program should be able to deal the cards for the Player and the Dealer in random order.
The deck will contain only the number cards and not the “face” cards. I.e. there is no Ace, King, Queen or Jack. Therefore, there will be a total of 36 cards in the deck.
First, the Dealer and the player will be dealt 1 card each.
Subsequently, the Player should get the option to “stand” or “hit” an additional card.
The Dealer will take additional cards till their total reaches 17 or higher.
If the Player total exceeds 21, the Player loses the round. If the Dealer total exceeds 21, the Player wins the round.
If the Player total is exactly 21, the Player wins the round